separating ui from main file and adding more tests
